Rename persistence -> config.
