+++ internal/config/gcs_test.go
@@ -1,4 +1,4 @@
-package persistence
+package config
 
 import (
 	"bytes"
